refactor: extract src/rg.rs + IgnoreMatcher + SOURCE_EXTENSIONS
Move all ripgrep helper functions and related types out of the
monolithic indexer.rs into a focused src/rg.rs module:

Items extracted from src/indexer.rs:
- IgnoreMatcher struct + impl (glob-based ignore-pattern filtering)
- SOURCE_EXTENSIONS constant
- walk_to_git_root (private helper for effective_rg_root)
- effective_rg_root (pub(crate))
- rg_find_definition (pub(crate))
- rg_find_references (pub)
- rg_find_implementors (pub)
- parse_rg_line (pub(crate))
- Private helpers: regex_escape (now pub(crate)), rg_raw,
  rg_files_with_matches, rg_word_in_files,
  parse_rg_line_with_content_rooted

Item moved from src/resolver.rs:
- build_rg_pattern (pub(crate)) — pure string function with no
  dependency on Indexer or FileData

Tests moved from src/indexer.rs to src/rg_tests.rs (referenced via
#[cfg(test)] #[path = "rg_tests.rs"] mod tests):
- rg_line_absolute_path_parsed / rg_line_relative_path_ignored
- refs_inner_class_does_not_bleed_across_contracts
- refs_inner_class_resolved_from_import_in_reference_file
- refs_decl_files_filtered_by_enclosing_class
- rg_find_definition_filters_ignored_dirs
- rg_find_references_filters_ignored_dirs

Backward-compat re-exports added to src/indexer.rs:
- pub(crate) use crate::rg::IgnoreMatcher  (callers unchanged)
- pub use crate::rg::SOURCE_EXTENSIONS     (callers unchanged)

Callers updated to use crate::rg:: directly:
- src/resolver.rs: parse_rg_line, rg_find_definition, build_rg_pattern,
  SOURCE_EXTENSIONS, IgnoreMatcher
- src/backend.rs: effective_rg_root, rg_find_definition,
  rg_find_references, rg_find_implementors
- src/indexer.rs internal callers: crate::rg::rg_find_definition (x2)

All 346 + 2 swift tests pass (cargo test).
