+++ src/indexer.rs
@@ -24,121 +24,10 @@ mod progress {
 use crate::parser;
 use crate::types::{FileData, SymbolEntry, FileIndexResult, WorkspaceIndexResult, IndexStats};
 
-// ─── Ignore pattern matcher ───────────────────────────────────────────────────
-
-/// Compiled workspace-level ignore patterns from `initializationOptions`.
-///
-/// Patterns follow gitignore glob semantics:
-/// - A bare pattern with no `/` (e.g. `bazel-*`) matches at any depth.
-/// - A pattern containing `/` (e.g. `build/**`) matches relative to the workspace root.
-/// - Absolute paths under the workspace root are normalized to relative before matching.
-pub(crate) struct IgnoreMatcher {
-    /// Original patterns as provided by the client (passed to `fd --exclude` as-is).
-    pub patterns: Vec<String>,
-    /// Arc-wrapped so the compiled set can be shared into `filter_entry` closures.
-    glob_set: Arc<globset::GlobSet>,
-    /// Workspace root this matcher was built for — used to relativize absolute paths.
-    root: std::path::PathBuf,
-}
-
-impl IgnoreMatcher {
-    /// Build an `IgnoreMatcher` from raw client patterns against `root`.
-    pub fn new(patterns: Vec<String>, root: &Path) -> Self {
-        let mut builder = globset::GlobSetBuilder::new();
-        for pat in &patterns {
-            // Normalize absolute paths that fall under the workspace root.
-            let normalized = if Path::new(pat.as_str()).is_absolute() {
-                match Path::new(pat.as_str()).strip_prefix(root) {
-                    Ok(rel) => rel.to_string_lossy().into_owned(),
-                    Err(_) => {
-                        log::warn!("ignorePatterns: absolute path {:?} is not under workspace root, skipping", pat);
-                        continue;
-                    }
-                }
-            } else {
-                pat.clone()
-            };
-
-            // Bare patterns (no `/`) match at any depth.
-            // Compile two variants:
-            //   `**/pattern`    — matches the directory entry itself (used in walkdir filter_entry)
-            //   `**/pattern/**` — matches all files inside a matching directory (used in filter_locs)
-            let glob_pats: Vec<String> = if !normalized.contains('/') {
-                vec![
-                    format!("**/{}", normalized),
-                    format!("**/{}/", normalized),   // trailing / for dir match
-                    format!("**/{normalized}/**"),
-                ]
-            } else {
-                vec![normalized]
-            };
-
-            for glob_pat in glob_pats {
-                match globset::Glob::new(&glob_pat) {
-                    Ok(g) => { builder.add(g); }
-                    Err(e) => { log::warn!("ignorePatterns: invalid pattern {:?}: {}", pat, e); }
-                }
-            }
-        }
-        let glob_set = builder.build().unwrap_or_else(|e| {
-            log::warn!("ignorePatterns: failed to build glob set: {}", e);
-            globset::GlobSetBuilder::new().build().unwrap()
-        });
-        Self { patterns, glob_set: Arc::new(glob_set), root: root.to_path_buf() }
-    }
-
-    pub fn is_empty(&self) -> bool {
-        self.patterns.is_empty()
-    }
-
-    /// Returns `true` if `rel_path` (relative to workspace root) should be excluded.
-    pub fn matches(&self, rel_path: &Path) -> bool {
-        self.glob_set.is_match(rel_path)
-    }
-
-    /// Clone the Arc-wrapped glob set for use in `filter_entry` closures.
-    pub fn glob_set(&self) -> Arc<globset::GlobSet> {
-        Arc::clone(&self.glob_set)
-    }
-
-    /// Remove locations whose file path is inside an ignored directory.
-    /// Paths are relativized against the workspace root this matcher was built for.
-    pub fn filter_locs(&self, locs: Vec<Location>) -> Vec<Location> {
-        locs.into_iter()
-            .filter(|loc| {
-                if let Ok(path) = loc.uri.to_file_path() {
-                    let rel = path.strip_prefix(&self.root).unwrap_or(&path);
-                    !self.matches(rel)
-                } else {
-                    true
-                }
-            })
-            .collect()
-    }
-
-    /// Remove file paths (absolute strings) that fall inside an ignored directory.
-    pub fn filter_file_strings(&self, files: Vec<String>) -> Vec<String> {
-        files
-            .into_iter()
-            .filter(|f| {
-                let path = Path::new(f);
-                let rel = path.strip_prefix(&self.root).unwrap_or(path);
-                !self.matches(rel)
-            })
-            .collect()
-    }
-
-    /// Remove `PathBuf` entries that fall inside an ignored directory.
-    pub fn filter_paths(&self, paths: Vec<std::path::PathBuf>) -> Vec<std::path::PathBuf> {
-        paths
-            .into_iter()
-            .filter(|p| {
-                let rel = p.strip_prefix(&self.root).unwrap_or(p);
-                !self.matches(rel)
-            })
-            .collect()
-    }
-}
+// Re-export rg-module items that existing callers reach via `crate::indexer::`.
+pub(crate) use crate::rg::IgnoreMatcher;
+pub use crate::rg::SOURCE_EXTENSIONS;
+use crate::rg::regex_escape;
 
 // ─── RAII guard for indexing_in_progress flag ─────────────────────────────────
 
@@ -153,9 +42,6 @@ impl Drop for IndexingGuard {
         log::debug!("IndexingGuard: cleared indexing_in_progress flag");
     }
 }
-
-/// Supported file extensions for indexing and rg/fd searches.
-pub const SOURCE_EXTENSIONS: &[&str] = &["kt", "java", "swift"];
 
 // ─── Indexing status file ─────────────────────────────────────────────────────
 
@@ -2632,18 +2518,6 @@ fn hash_str(s: &str) -> u64 {
 
 // ─── rg cross-file fallback ──────────────────────────────────────────────────
 
-/// Walk up from `file` until a `.git` directory is found, returning that
-/// ancestor as the project root.  Returns `None` if no `.git` is found.
-fn walk_to_git_root(file: &Path) -> Option<std::path::PathBuf> {
-    let mut cur = file.parent()?;
-    loop {
-        if cur.join(".git").exists() {
-            return Some(cur.to_path_buf());
-        }
-        cur = cur.parent()?;
-    }
-}
-
 /// Given a list of type names (e.g. `["MviViewModel", "ViewModel"]`), find the
 /// source files in `root` that declare them using a fast `rg` search.
 /// Returns deduplicated, existing paths (at most one file per type).
@@ -2672,383 +2546,6 @@ fn find_files_for_types(names: &[String], root: &Path, matcher: Option<&IgnoreMa
         .filter(|p| p.exists())
         .collect();
     matcher.map_or(paths.clone(), |m| m.filter_paths(paths))
-}
-
-/// Return the best search root for rg/fd fallbacks given:
-/// - `workspace_root` — the globally configured root (may point to a different project)
-/// - `open_file`      — the file the user has open right now
-///
-/// If `open_file` lives inside `workspace_root`, use `workspace_root`.
-/// Otherwise walk up from `open_file` to find a `.git` root and use that,
-/// so rg searches the *actual* project even when the workspace config is stale.
-pub(crate) fn effective_rg_root(
-    workspace_root: Option<&Path>,
-    open_file:      Option<&Path>,
-) -> Option<std::path::PathBuf> {
-    match (workspace_root, open_file) {
-        (Some(root), Some(fp)) if fp.starts_with(root) => Some(root.to_path_buf()),
-        (_, Some(fp)) => walk_to_git_root(fp)
-            .or_else(|| fp.parent().map(|p| p.to_path_buf()))
-            .or_else(|| std::env::current_dir().ok()),
-        (Some(root), None) => Some(root.to_path_buf()),
-        (None, None) => std::env::current_dir().ok(),
-    }
-}
-
-/// Run `rg` to find definition sites for `name`, scoped to `root`.
-///
-/// When `root` is an absolute path, rg outputs absolute paths in results.
-/// Passing workspace root here is essential; without it rg would search
-/// from CWD which may not be the project when spawned by the editor.
-///
-/// Results in directories matched by `matcher` are filtered out.
-pub(crate) fn rg_find_definition(name: &str, root: Option<&Path>, matcher: Option<&IgnoreMatcher>) -> Vec<Location> {
-    let pattern = crate::resolver::build_rg_pattern(name);
-
-    // Use the provided root, or fall back to CWD (which editors like Helix
-    // set to the workspace root when spawning the LSP server).
-    let search_root: std::borrow::Cow<Path> = match root {
-        Some(r) => std::borrow::Cow::Borrowed(r),
-        None    => std::borrow::Cow::Owned(std::env::current_dir().unwrap_or_default()),
-    };
-
-    let mut cmd = Command::new("rg");
-    cmd.args([
-        "--no-heading",
-        "--with-filename",
-        "--line-number",
-        "--column",
-        // NOTE: rg has no --absolute-path flag; absolute output comes from
-        // passing an absolute search root as the positional argument.
-    ]);
-    for ext in SOURCE_EXTENSIONS {
-        cmd.args(["--glob", &format!("*.{ext}")]);
-    }
-    cmd.args(["-e", &pattern]);
-    cmd.arg(search_root.as_ref());
-
-    let out = match cmd.output() {
-        Ok(o) if o.status.success() => o,
-        _ => return vec![],
-    };
-
-    let locs: Vec<Location> = String::from_utf8_lossy(&out.stdout)
-        .lines()
-        .filter_map(|l| parse_rg_line_with_content_rooted(l, &search_root).map(|(loc, _)| loc))
-        .collect();
-
-    matcher.map_or(locs.clone(), |m| m.filter_locs(locs))
-}
-
-/// Run `rg` to find all *usages* of `name` in the project.
-///
-/// Uses `--word-regexp` so only whole-word matches are returned.
-/// If `include_decl` is false, declaration lines are filtered out by
-/// excluding lines that contain declaration keywords before `name`.
-/// If `from_uri` is provided, the source file is excluded when
-/// `include_decl` is false (the definition is already known).
-///
-/// Results in directories matched by `matcher` are filtered out.
-pub fn rg_find_references(
-    name:         &str,
-    parent_class: Option<&str>,
-    declared_pkg: Option<&str>,
-    root:         Option<&Path>,
-    include_decl: bool,
-    from_uri:     &Url,
-    // Absolute file paths where `name` is declared — always included in bare-word
-    // search so the declaration site itself is never missed (it uses bare `Name`,
-    // not the qualified `Parent.Name` form that Pass A searches for).
-    decl_files:   &[String],
-    matcher:      Option<&IgnoreMatcher>,
-) -> Vec<Location> {
-    let search_root: std::borrow::Cow<Path> = match root {
-        Some(r) => std::borrow::Cow::Borrowed(r),
-        None    => std::borrow::Cow::Owned(std::env::current_dir().unwrap_or_default()),
-    };
-
-    let safe_name: String = regex_escape(name);
-    let decl_kws = ["class ", "interface ", "object ", "fun ", "val ", "var ",
-                    "typealias ", "enum class ", "enum ",
-                    // Swift
-                    "struct ", "protocol ", "func ", "let ", "extension "];
-
-    let filter = |(loc, content): (Location, String)| -> Option<Location> {
-        let trimmed = content.trim_start();
-        // Import and package lines are never real references.
-        if trimmed.starts_with("import ") || trimmed.starts_with("package ") {
-            return None;
-        }
-        if !include_decl {
-            let is_decl = decl_kws.iter().any(|kw| content.contains(kw))
-                && loc.uri.as_str() == from_uri.as_str();
-            if is_decl { return None; }
-        }
-        Some(loc)
-    };
-
-    let result = if let Some(parent) = parent_class {
-        // ── Scoped references: parent class is known ──────────────────────────
-        //
-        // Pass A: qualified form `ParentClass.Name` — works in any file.
-        let safe_parent = regex_escape(parent);
-        let qualified_pat = format!(r"\b{}\.\b{}\b", safe_parent, safe_name);
-        let mut locs: Vec<Location> = rg_raw(&qualified_pat, &search_root)
-            .into_iter()
-            .filter_map(filter)
-            .collect();
-        eprintln!("[refs] parent={parent:?} Pass-A qualified={} locs", locs.len());
-
-        // Pass B: bare `Name` restricted to files that directly import the inner
-        // class itself (`import …ParentClass.Name` or `import …ParentClass.*`)
-        // OR are in the same package.
-        //
-        // NOTE: we intentionally do NOT match files that only import the parent
-        // class itself (`import …ParentClass`) — those files use the qualified
-        // form `ParentClass.Name` which is already captured by Pass A, and
-        // including them causes massive false-positive counts (e.g. every
-        // ViewModel importing another ViewModel that also has a sealed `Effect`).
-        //
-        // Step B1 — files with explicit inner-class import.
-        // Pattern must match the parent and name as ADJACENT dot-segments:
-        //   import …ParentClass.Name   or   import …ParentClass.*
-        // NOT files that merely mention both words (e.g. OtherContract.State).
-        let direct_import_pat = format!(
-            r"import[^\n]*\b{}\.({}|\*)\b",
-            safe_parent, safe_name
-        );
-        let candidate_files = rg_files_with_matches(&direct_import_pat, &search_root);
-        // Filter candidate files against ignore patterns before searching them.
-        let candidate_files = matcher
-            .map_or(candidate_files.clone(), |m| m.filter_file_strings(candidate_files));
-
-        // Step B2 — files in the same package as the parent class declaration.
-        // NOTE: for inner classes, same-package files use the QUALIFIED form
-        // `ParentClass.Name` which is already caught by Pass A. Adding them to
-        // the bare-name search causes false positives (e.g. AbilitiesSectionViewModel
-        // in the same package has its own `State`). So we skip same-package here.
-
-        // Merge candidate file sets.
-        // Always include declaration files so the declaration site itself is
-        // never missed (it uses bare `Name`, not the qualified `Parent.Name` form).
-        let mut all_files: Vec<String> = candidate_files;
-        for f in decl_files {
-            if !all_files.contains(f) { all_files.push(f.clone()); }
-        }
-
-        if !all_files.is_empty() {
-            let bare_hits = rg_word_in_files(&safe_name, &all_files);
-            eprintln!("[refs] Pass-B candidate files={} bare_hits={}", all_files.len(), bare_hits.len());
-            for (loc, content) in bare_hits {
-                if let Some(loc) = filter((loc, content)) {
-                    // Deduplicate against the qualified hits.
-                    if !locs.iter().any(|l: &Location| l.uri == loc.uri && l.range.start == loc.range.start) {
-                        locs.push(loc);
-                    }
-                }
-            }
-        }
-
-        locs
-    } else if let Some(dpkg) = declared_pkg {
-        // ── Top-level symbol with known declared package ──────────────────────
-        // Only search files that import `declared_pkg.Name` or `declared_pkg.*`
-        // or are in the same package. This avoids the "13000 matches for Effect"
-        // problem where every ViewModel has an inner class with the same name.
-        let safe_pkg = regex_escape(dpkg);
-        let import_pat = format!(
-            r"import[^\n]*\b{safe_pkg}\b[^\n]*\b{safe_name}\b|import[^\n]*\b{safe_pkg}\b\.\*"
-        );
-        let pkg_pat = format!(r"^\s*package\s+{safe_pkg}\s*$");
-
-        let mut candidate_files = rg_files_with_matches(&import_pat, &search_root);
-        for f in rg_files_with_matches(&pkg_pat, &search_root) {
-            if !candidate_files.contains(&f) { candidate_files.push(f); }
-        }
-        // Filter candidate files against ignore patterns before searching them.
-        let candidate_files = matcher
-            .map_or(candidate_files.clone(), |m| m.filter_file_strings(candidate_files));
-
-        if candidate_files.is_empty() {
-            return vec![];
-        }
-        rg_word_in_files(&safe_name, &candidate_files)
-            .into_iter()
-            .filter_map(filter)
-            .collect()
-    } else {
-        // ── Fully unscoped: lowercase / unknown symbol ────────────────────────
-        let mut cmd = Command::new("rg");
-        cmd.args([
-            "--no-heading", "--with-filename", "--line-number", "--column",
-            "--word-regexp",
-        ]);
-        for ext in SOURCE_EXTENSIONS {
-            cmd.args(["--glob", &format!("*.{ext}")]);
-        }
-        cmd.args(["-e", &safe_name]);
-        cmd.arg(search_root.as_ref());
-
-        let out = match cmd.output() {
-            Ok(o) if !o.stdout.is_empty() => o,
-            _ => return vec![],
-        };
-
-        String::from_utf8_lossy(&out.stdout)
-            .lines()
-            .filter_map(|l| parse_rg_line_with_content_rooted(l, &search_root))
-            .filter_map(filter)
-            .collect()
-    };
-
-    if let Some(m) = matcher { m.filter_locs(result) } else { result }
-}
-
-fn regex_escape(s: &str) -> String {
-    s.chars().flat_map(|c| {
-        if c.is_alphanumeric() || c == '_' { vec![c] } else { vec!['\\', c] }
-    }).collect()
-}
-
-/// Run rg with a regex pattern; return `(Location, line_content)` pairs.
-fn rg_raw(pattern: &str, root: &Path) -> Vec<(Location, String)> {
-    let mut cmd = Command::new("rg");
-    cmd.args(["--no-heading", "--with-filename", "--line-number", "--column"]);
-    for ext in SOURCE_EXTENSIONS {
-        cmd.args(["--glob", &format!("*.{ext}")]);
-    }
-    cmd.args(["-e", pattern]).arg(root);
-    let out = match cmd.output() {
-        Ok(o) if !o.stdout.is_empty() => o,
-        _ => return vec![],
-    };
-    String::from_utf8_lossy(&out.stdout)
-        .lines()
-        .filter_map(|l| parse_rg_line_with_content_rooted(l, root))
-        .collect()
-}
-
-/// Run `rg -l` to get the list of files matching a pattern.
-fn rg_files_with_matches(pattern: &str, root: &Path) -> Vec<String> {
-    let mut cmd = Command::new("rg");
-    cmd.arg("-l");
-    for ext in SOURCE_EXTENSIONS {
-        cmd.args(["--glob", &format!("*.{ext}")]);
-    }
-    cmd.args(["-e", pattern]).arg(root);
-    let out = match cmd.output() {
-        Ok(o) if !o.stdout.is_empty() => o,
-        _ => return vec![],
-    };
-    String::from_utf8_lossy(&out.stdout)
-        .lines()
-        .map(|l| {
-            let p = std::path::Path::new(l);
-            if p.is_absolute() {
-                l.to_owned()
-            } else {
-                root.join(l).to_string_lossy().into_owned()
-            }
-        })
-        .collect()
-}
-
-/// Run `rg --word-regexp NAME` restricted to specific files.
-fn rg_word_in_files(safe_name: &str, files: &[String]) -> Vec<(Location, String)> {
-    if files.is_empty() { return vec![]; }
-    let out = match Command::new("rg")
-        .args(["--no-heading", "--with-filename", "--line-number", "--column",
-               "--word-regexp", "-e", safe_name, "--"])
-        .args(files)
-        .output()
-    {
-        Ok(o) if !o.stdout.is_empty() => o,
-        _ => return vec![],
-    };
-    // Files passed to rg_word_in_files are already absolute (from rg_files_with_matches).
-    String::from_utf8_lossy(&out.stdout)
-        .lines()
-        .filter_map(|l| parse_rg_line_with_content_rooted(l, std::path::Path::new("/")))
-        .collect()
-}
-
-fn parse_rg_line_with_content_rooted(line: &str, root: &Path) -> Option<(Location, String)> {
-    let mut parts = line.splitn(4, ':');
-    let file     = parts.next()?;
-    let line_num: u32 = parts.next()?.trim().parse().ok()?;
-    let col:      u32 = parts.next()?.trim().parse().ok()?;
-    let content  = parts.next().unwrap_or("").to_string();
-
-    let path = std::path::Path::new(file);
-    let abs_path = if path.is_absolute() {
-        path.to_path_buf()
-    } else {
-        root.join(path)
-    };
-
-    let uri = Url::from_file_path(&abs_path).ok()?;
-    let pos = Position::new(line_num.saturating_sub(1), col.saturating_sub(1));
-    Some((Location { uri, range: Range::new(pos, pos) }, content))
-}
-
-/// Quick heuristic rg-based implementor finder. Scans files that mention `name`
-/// and returns locations where the line looks like a declaration/implementation
-/// of that type (Kotlin/Java `class Foo : Interface`, `implements`, Swift
-/// `class Foo: Protocol`, `struct Foo: Protocol`). This is a fallback when the
-/// subtype index is empty during cold indexing.
-///
-/// Results in directories matched by `matcher` are filtered out.
-pub fn rg_find_implementors(name: &str, root: Option<&Path>, matcher: Option<&IgnoreMatcher>) -> Vec<Location> {
-    let safe = name.to_string();
-    let root = match root {
-        Some(r) => r,
-        None => return vec![],
-    };
-    // Search for the name in source files.
-    let mut cmd = Command::new("rg");
-    cmd.args(["--no-heading", "--with-filename", "--line-number", "--column", "-e", &safe]);
-    for ext in SOURCE_EXTENSIONS { cmd.args(["--glob", &format!("*.{ext}")]); }
-    cmd.arg(root);
-    let out = match cmd.output() {
-        Ok(o) if !o.stdout.is_empty() => o,
-        _ => return vec![],
-    };
-    let locs: Vec<Location> = String::from_utf8_lossy(&out.stdout)
-        .lines()
-        .filter_map(|l| parse_rg_line_with_content_rooted(l, root))
-        .filter_map(|(loc, content)| {
-            let line = content.trim();
-            // Heuristics: declaration-like lines
-            // Kotlin/Java: class Foo, interface Foo, enum class Foo, class Foo : Interface
-            // Java implements: class Foo implements Interface
-            // Swift: class Foo: Protocol, struct Foo: Protocol, extension Foo: Protocol
-            let lower = line.to_lowercase();
-            if lower.contains("class ") || lower.contains("struct ") || lower.contains("interface") || lower.contains("enum") || lower.contains("extension ") {
-                // Check that the name appears as a word and near a declaration keyword
-                if line.contains(name) {
-                    return Some(loc);
-                }
-            }
-            None
-        })
-        .collect();
-    matcher.map_or(locs.clone(), |m| m.filter_locs(locs))
-}
-
-pub(crate) fn parse_rg_line(line: &str) -> Option<Location> {
-    // format: /abs/path/to/File.kt:line:col:content
-    let mut parts = line.splitn(4, ':');
-    let file     = parts.next()?;
-    let line_num: u32 = parts.next()?.trim().parse().ok()?;
-    let col:      u32 = parts.next()?.trim().parse().ok()?;
-
-    let path = std::path::Path::new(file);
-    // Silently skip if rg somehow gave us a relative path.
-    if !path.is_absolute() { return None; }
-
-    let uri = Url::from_file_path(path).ok()?;
-    let pos = Position::new(line_num.saturating_sub(1), col.saturating_sub(1));
-    Some(Location { uri, range: Range::new(pos, pos) })
 }
 
 // ─── misc helpers ────────────────────────────────────────────────────────────
@@ -3641,7 +3138,7 @@ fn lambda_receiver_type_named_arg_ml(
                     // On-demand: use rg to find and index the outer class.
                     let root = idx.workspace_root.read().unwrap().clone();
                     let matcher = idx.ignore_matcher.read().unwrap().clone();
-                    let rg_locs = rg_find_definition(
+                    let rg_locs = crate::rg::rg_find_definition(
                         outer, root.as_deref(), matcher.as_deref()
                     );
                     for loc in &rg_locs {
@@ -4020,7 +3517,7 @@ fn find_fun_signature_full(fn_name: &str, idx: &Indexer, uri: &Url) -> Option<St
     // Slow path: rg to locate the definition, index on-demand.
     let root = idx.workspace_root.read().unwrap().clone();
     let matcher = idx.ignore_matcher.read().unwrap().clone();
-    let locs = rg_find_definition(fn_name, root.as_deref(), matcher.as_deref());
+    let locs = crate::rg::rg_find_definition(fn_name, root.as_deref(), matcher.as_deref());
     for loc in &locs {
         let file_uri_str = loc.uri.as_str();
         if !idx.files.contains_key(file_uri_str) {
@@ -5159,24 +4656,6 @@ class Child : Base
         let (u, idx) = indexed("/t.kt", "package com.example\nclass Foo");
         let uris = idx.packages.get("com.example").unwrap();
         assert!(uris.contains(&u.to_string()));
-    }
-
-    // ── parse_rg_line ────────────────────────────────────────────────────────
-
-    #[test]
-    fn rg_line_absolute_path_parsed() {
-        let line = "/home/user/project/Foo.kt:10:5:class Foo {";
-        let loc = parse_rg_line(line).unwrap();
-        assert_eq!(loc.range.start.line,      9); // 1-indexed → 0-indexed
-        assert_eq!(loc.range.start.character, 4);
-        assert_eq!(loc.uri.path(), "/home/user/project/Foo.kt");
-    }
-
-    #[test]
-    fn rg_line_relative_path_ignored() {
-        // Before the fix this would panic / produce a wrong URI
-        let line = "src/Foo.kt:10:5:class Foo {";
-        assert!(parse_rg_line(line).is_none(), "relative paths must be ignored");
     }
 
     // ── parse_count: verify deduplication ───────────────────────────────────
@@ -6324,220 +5803,6 @@ class State
             "it inside multi-param lambda should be None, got: {result:?}");
     }
 
-    // ── rg_find_references scoping ───────────────────────────────────────────
-
-    /// Write `content` to `dir/rel_path` and return the absolute path as String.
-    fn write_temp(dir: &std::path::Path, rel_path: &str, content: &str) -> String {
-        let p = dir.join(rel_path);
-        if let Some(parent) = p.parent() { std::fs::create_dir_all(parent).unwrap(); }
-        std::fs::write(&p, content).unwrap();
-        p.to_str().unwrap().to_owned()
-    }
-
-    /// `rg_find_references` must not bleed references across sealed interfaces
-    /// that share the same inner name (`Event`) but belong to different contracts.
-    ///
-    /// Layout:
-    ///   activate_contract.kt   — declares interface ActivateUpdateAppContract { sealed interface Event }
-    ///   other_contract.kt      — declares interface OtherContract             { sealed interface Event }
-    ///   activate_vm.kt         — imports ActivateUpdateAppContract.Event, uses bare `Event`
-    ///   other_vm.kt            — imports OtherContract.Event,             uses bare `Event`
-    ///
-    /// Finding refs for ActivateUpdateAppContract.Event must return hits in
-    /// activate_contract.kt and activate_vm.kt ONLY — not other_vm.kt.
-    #[test]
-    fn refs_inner_class_does_not_bleed_across_contracts() {
-        let dir = tempfile::tempdir().expect("tempdir");
-        let root = dir.path();
-
-        write_temp(root, "activate_contract.kt", concat!(
-            "package com.example.activate\n",
-            "interface ActivateUpdateAppContract {\n",
-            "  sealed interface Event\n",
-            "}\n",
-        ));
-        write_temp(root, "other_contract.kt", concat!(
-            "package com.example.other\n",
-            "interface OtherContract {\n",
-            "  sealed interface Event\n",
-            "}\n",
-        ));
-        write_temp(root, "activate_vm.kt", concat!(
-            "package com.example.activate\n",
-            "import com.example.activate.ActivateUpdateAppContract.Event\n",
-            "class ActivateViewModel {\n",
-            "  fun handle(e: Event) {}\n",
-            "}\n",
-        ));
-        write_temp(root, "other_vm.kt", concat!(
-            "package com.example.other\n",
-            "import com.example.other.OtherContract.Event\n",
-            "class OtherViewModel {\n",
-            "  fun handle(e: Event) {}\n",
-            "}\n",
-        ));
-
-        let activate_uri = Url::from_file_path(root.join("activate_contract.kt")).unwrap();
-        let activate_decl = root.join("activate_contract.kt").to_str().unwrap().to_owned();
-
-        // Simulate: cursor on declaration of Event inside ActivateUpdateAppContract.
-        // parent_class = "ActivateUpdateAppContract", declared_pkg = "com.example.activate"
-        let locs = super::rg_find_references(
-            "Event",
-            Some("ActivateUpdateAppContract"),
-            Some("com.example.activate"),   // declared_pkg
-            Some(root),
-            true,  // include_declaration
-            &activate_uri,
-            &[activate_decl],
-            None,  // no ignore patterns in this test
-        );
-
-        let hit_files: std::collections::HashSet<String> = locs.iter()
-            .map(|l| l.uri.to_file_path().unwrap().file_name().unwrap().to_str().unwrap().to_owned())
-            .collect();
-
-        assert!(hit_files.contains("activate_contract.kt"),
-            "should include declaration file; got: {hit_files:?}");
-        assert!(hit_files.contains("activate_vm.kt"),
-            "should include file that imports ActivateUpdateAppContract.Event; got: {hit_files:?}");
-        assert!(!hit_files.contains("other_vm.kt"),
-            "must NOT include file that only imports OtherContract.Event; got: {hit_files:?}");
-        assert!(!hit_files.contains("other_contract.kt"),
-            "must NOT include OtherContract declaration; got: {hit_files:?}");
-    }
-
-    /// When cursor is on `Event` inside a file that imports `OtherContract.Event`,
-    /// refs must not include files that only import `ActivateUpdateAppContract.Event`.
-    #[test]
-    fn refs_inner_class_resolved_from_import_in_reference_file() {
-        let dir = tempfile::tempdir().expect("tempdir");
-        let root = dir.path();
-
-        write_temp(root, "activate_contract.kt", concat!(
-            "package com.example.activate\n",
-            "interface ActivateUpdateAppContract {\n",
-            "  sealed interface Event\n",
-            "}\n",
-        ));
-        write_temp(root, "other_contract.kt", concat!(
-            "package com.example.other\n",
-            "interface OtherContract {\n",
-            "  sealed interface Event\n",
-            "}\n",
-        ));
-        write_temp(root, "activate_vm.kt", concat!(
-            "package com.example.activate\n",
-            "import com.example.activate.ActivateUpdateAppContract.Event\n",
-            "class ActivateViewModel {\n",
-            "  fun handle(e: Event) {}\n",
-            "}\n",
-        ));
-        write_temp(root, "other_vm.kt", concat!(
-            "package com.example.other\n",
-            "import com.example.other.OtherContract.Event\n",
-            "class OtherViewModel {\n",
-            "  fun handle(e: Event) {}\n",
-            "}\n",
-        ));
-
-        // Simulate: cursor on `Event` inside other_vm.kt (a reference, not declaration).
-        // resolve_symbol_via_import on other_vm.kt → parent=OtherContract, pkg=com.example.other
-        let other_vm_uri = Url::from_file_path(root.join("other_vm.kt")).unwrap();
-        let other_decl = root.join("other_contract.kt").to_str().unwrap().to_owned();
-
-        let locs = super::rg_find_references(
-            "Event",
-            Some("OtherContract"),
-            Some("com.example.other"),
-            Some(root),
-            true,
-            &other_vm_uri,
-            &[other_decl],
-            None,
-        );
-
-        let hit_files: std::collections::HashSet<String> = locs.iter()
-            .map(|l| l.uri.to_file_path().unwrap().file_name().unwrap().to_str().unwrap().to_owned())
-            .collect();
-
-        assert!(hit_files.contains("other_contract.kt"),
-            "should include OtherContract declaration; got: {hit_files:?}");
-        assert!(hit_files.contains("other_vm.kt"),
-            "should include file importing OtherContract.Event; got: {hit_files:?}");
-        assert!(!hit_files.contains("activate_vm.kt"),
-            "must NOT include file importing ActivateUpdateAppContract.Event; got: {hit_files:?}");
-    }
-
-    /// Regression: when `decl_files` is unfiltered it includes ALL contracts that
-    /// declare a `sealed interface Event`, causing every consumer ViewModel to appear
-    /// in results for an unrelated contract's Event.
-    ///
-    /// Layout: two contracts each with `sealed interface Event`, two ViewModels each
-    /// importing their own contract's Event.  Finding refs for DashboardContract.Event
-    /// must NOT return VisitBranchViewModel even though both are in `decl_files` when
-    /// unfiltered by enclosing-class.
-    #[test]
-    fn refs_decl_files_filtered_by_enclosing_class() {
-        let dir = tempfile::tempdir().expect("tempdir");
-        let root = dir.path();
-
-        write_temp(root, "DashboardContract.kt", concat!(
-            "package com.example.dashboard\n",
-            "interface DashboardContract {\n",
-            "  sealed interface Event\n",
-            "}\n",
-        ));
-        write_temp(root, "VisitBranchContract.kt", concat!(
-            "package com.example.visitbranch\n",
-            "interface VisitBranchContract {\n",
-            "  sealed interface Event\n",
-            "}\n",
-        ));
-        write_temp(root, "DashboardViewModel.kt", concat!(
-            "package com.example.dashboard\n",
-            "import com.example.dashboard.DashboardContract.Event\n",
-            "class DashboardViewModel {\n",
-            "  fun handle(e: Event) {}\n",
-            "}\n",
-        ));
-        write_temp(root, "VisitBranchViewModel.kt", concat!(
-            "package com.example.visitbranch\n",
-            "import com.example.visitbranch.VisitBranchContract.Event\n",
-            "class VisitBranchViewModel {\n",
-            "  fun handle(e: Event) {}\n",
-            "}\n",
-        ));
-
-        let dashboard_uri = Url::from_file_path(root.join("DashboardContract.kt")).unwrap();
-        // decl_files filtered to only DashboardContract.kt (enclosing = DashboardContract)
-        let dashboard_decl = root.join("DashboardContract.kt").to_str().unwrap().to_owned();
-
-        let locs = super::rg_find_references(
-            "Event",
-            Some("DashboardContract"),
-            Some("com.example.dashboard"),
-            Some(root),
-            true,
-            &dashboard_uri,
-            &[dashboard_decl],  // NOT including VisitBranchContract.kt
-            None,
-        );
-
-        let hit_files: std::collections::HashSet<String> = locs.iter()
-            .map(|l| l.uri.to_file_path().unwrap().file_name().unwrap().to_str().unwrap().to_owned())
-            .collect();
-
-        assert!(hit_files.contains("DashboardContract.kt"),
-            "should include DashboardContract declaration; got: {hit_files:?}");
-        assert!(hit_files.contains("DashboardViewModel.kt"),
-            "should include DashboardViewModel; got: {hit_files:?}");
-        assert!(!hit_files.contains("VisitBranchViewModel.kt"),
-            "must NOT include VisitBranchViewModel; got: {hit_files:?}");
-        assert!(!hit_files.contains("VisitBranchContract.kt"),
-            "must NOT include VisitBranchContract; got: {hit_files:?}");
-    }
-
     // ── subtypes index (goToImplementation) ──────────────────────────────────
 
     #[test]
@@ -7391,75 +6656,6 @@ class Foo @Inject constructor(
             !indexer.definitions.contains_key("VendorClass"),
             "VendorClass (under third-party/) must be excluded"
         );
-    }
-
-    /// `rg_find_definition` must not return results from ignored directories.
-    #[test]
-    fn rg_find_definition_filters_ignored_dirs() {
-        let dir = tempfile::TempDir::new().expect("create tempdir");
-        let root = dir.path();
-
-        std::fs::create_dir_all(root.join("src")).unwrap();
-        std::fs::write(root.join("src/Real.kt"),
-            "package com.example\nclass MyClass\n"
-        ).unwrap();
-
-        std::fs::create_dir_all(root.join("buildSrc/generated")).unwrap();
-        std::fs::write(root.join("buildSrc/generated/MyClass.kt"),
-            "package com.example\nclass MyClass\n"
-        ).unwrap();
-
-        let matcher = IgnoreMatcher::new(vec!["buildSrc".to_owned()], root);
-        let locs = super::rg_find_definition("MyClass", Some(root), Some(&matcher));
-        let files: Vec<String> = locs.iter()
-            .map(|l| l.uri.to_file_path().unwrap().to_string_lossy().into_owned())
-            .collect();
-
-        assert!(files.iter().any(|f| f.contains("src/Real.kt")),
-            "must include real source; got: {files:?}");
-        assert!(!files.iter().any(|f| f.contains("buildSrc")),
-            "must not include buildSrc results; got: {files:?}");
-    }
-
-    /// `rg_find_references` must exclude candidate files from ignored directories.
-    #[test]
-    fn rg_find_references_filters_ignored_dirs() {
-        let dir = tempfile::TempDir::new().expect("create tempdir");
-        let root = dir.path();
-
-        std::fs::create_dir_all(root.join("src")).unwrap();
-        std::fs::write(root.join("src/Contract.kt"),
-            "package com.example\nclass Contract {\n  class Event\n}\n"
-        ).unwrap();
-        std::fs::write(root.join("src/User.kt"),
-            "package com.example\nimport com.example.Contract.Event\nfun use(e: Event) {}\n"
-        ).unwrap();
-
-        std::fs::create_dir_all(root.join("buildSrc")).unwrap();
-        std::fs::write(root.join("buildSrc/Contract.kt"),
-            "package com.example\nclass Contract {\n  class Event\n}\n"
-        ).unwrap();
-
-        let uri = Url::from_file_path(root.join("src/Contract.kt")).unwrap();
-        let decl = root.join("src/Contract.kt").to_str().unwrap().to_owned();
-        let matcher = IgnoreMatcher::new(vec!["buildSrc".to_owned()], root);
-
-        let locs = super::rg_find_references(
-            "Event",
-            Some("Contract"),
-            Some("com.example"),
-            Some(root),
-            true,
-            &uri,
-            &[decl],
-            Some(&matcher),
-        );
-        let files: Vec<String> = locs.iter()
-            .map(|l| l.uri.to_file_path().unwrap().to_string_lossy().into_owned())
-            .collect();
-
-        assert!(!files.iter().any(|f| f.contains("buildSrc")),
-            "must not include buildSrc in references; got: {files:?}");
     }
 }
